feat(main): Add pre-tasks and post-tasks to subcommand_handler
This commit adds the `pre_tasks` and `post_tasks` functions to the `subcommand_handler` function in the `main.rs` file. These tasks are executed before and after processing the content of input files, respectively. Additionally, the code has been updated to remove unnecessary parameters from the `subcommand_handler` function.

- Added import statements for `chewbekka::pre_tasks` and `chewbekka::post_tasks`
- Updated function signature of `subcommand_handler` to remove unnecessary parameters
- Added calls to `pre_tasks` and `post_tasks`
- Removed unnecessary parameter from call to `write_md_file`

Note: The commit message does not include a detailed description as it is not provided in the code changes.